requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
sqlalchemy==2.0.35
alembic==1.13.2
celery==5.4.0
redis==5.1.1
boto3==1.35.23
python-multipart==0.0.9
pydantic-settings==2.5.2
httpx==0.27.2
psycopg2-binary==2.9.9
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
stripe==10.10.0
websockets==13.1
